ci: relocate some code
