+++ .dagger/main.go
@@ -1,9 +1,6 @@
 package main
 
 import (
-	"context"
-	"fmt"
-
 	"github.com/openmeterio/openmeter/.dagger/internal/dagger"
 )
 
@@ -25,87 +22,4 @@ func New(
 	return &Openmeter{
 		Source: source,
 	}
-}
-
-func (m *Openmeter) Ci(ctx context.Context) (*dagger.Directory, error) {
-	p := newPipeline(ctx)
-
-	trivy := dag.Trivy(dagger.TrivyOpts{
-		Cache:             cacheVolume("trivy"),
-		WarmDatabaseCache: true,
-	})
-
-	containerImages := m.Build().containerImages("ci")
-
-	helmChartOpenMeter := m.Build().helmChart("openmeter", "0.0.0").File()
-	helmChartBenthosCollector := m.Build().helmChart("benthos-collector", "0.0.0").File()
-	helmCharts := dag.Directory().WithFiles("", []*dagger.File{helmChartOpenMeter, helmChartBenthosCollector})
-
-	releaseAssets := dag.Directory().WithFiles("", m.releaseAssets("ci"))
-
-	generated := dag.Directory().
-		WithFile("", m.Generate().Openapi()).
-		WithDirectory("sdk/python", m.Generate().PythonSdk()).
-		WithDirectory("sdk/node", m.Generate().NodeSdk()).
-		WithDirectory("sdk/web", m.Generate().WebSdk())
-
-	dir := dag.Directory().
-		WithFile("scans/image.sarif", trivy.Container(containerImages[0]).Report("sarif")).
-		WithFile("scans/helm-openmeter.sarif", trivy.HelmChart(helmChartOpenMeter).Report("sarif")).
-		WithFile("scans/helm-benthos-collector.sarif", trivy.HelmChart(helmChartBenthosCollector).Report("sarif")).
-		WithDirectory("charts/", helmCharts).
-		WithDirectory("release/", releaseAssets).
-		WithDirectory("generated/", generated)
-
-	p.addJobs(
-		m.Generate().Check,
-
-		wrapSyncable(m.Test()),
-		m.Lint().All,
-
-		// TODO: version should be the commit hash (if any?)?
-		wrapSyncables(m.Build().containerImages("ci")),
-
-		wrapSyncable(dir),
-	)
-
-	return dir, p.wait()
-}
-
-func (m *Openmeter) Test() *dagger.Container {
-	// TODO: customize user and password
-	postgres := dag.Postgres(dagger.PostgresOpts{
-		Version: postgresVersion,
-	}).WithDatabase("svix")
-
-	svix := dag.Svix(dagger.SvixOpts{
-		Version:   svixVersion,
-		Postgres:  postgres.AsSvixPostgres(),
-		Database:  "svix",
-		JwtSecret: dag.SetSecret("svix-jwt-secret", SvixJWTSingingSecret),
-	})
-
-	return goModuleCross("").
-		WithSource(m.Source).
-		WithEnvVariable("POSTGRES_HOST", "postgres").
-		WithEnvVariable("SVIX_HOST", "svix").
-		WithEnvVariable("SVIX_JWT_SECRET", SvixJWTSingingSecret).
-		WithServiceBinding("postgres", postgres.Service()).
-		WithServiceBinding("svix", svix.Service()).
-		Exec([]string{"go", "test", "-tags", "musl", "-v", "./..."})
-}
-
-func (m *Openmeter) QuickstartTest(
-	service *dagger.Service,
-
-	// +default=8888
-	port int,
-) *dagger.Container {
-	return goModule().
-		WithModuleCache(cacheVolume("go-mod-quickstart")).
-		WithBuildCache(cacheVolume("go-build-quickstart")).
-		WithSource(m.Source).
-		WithEnvVariable("OPENMETER_ADDRESS", fmt.Sprintf("http://openmeter:%d", port)).
-		WithServiceBinding("openmeter", service).
-		Exec([]string{"go", "test", "-v", "./quickstart/"})
 }
